feat: Add loadable.types and loadable.const files
Relocate Loadable component-related types and constants into newly created
files, loadable.types and loadable.const, for enhanced project management.
