+++ components/loadable/loadable.const.ts
@@ -1,0 +1,5 @@
+export type SPINNER = (typeof SPINNER)[keyof typeof SPINNER];
+export const SPINNER = {
+  authForm: 'authForm',
+  verificationConfirm: 'verificationConfirm',
+} as const;
